requirements.txt
django-phonenumber-field
phonenumbers
email_validator
django-crispy-forms==1.8.1
Pillow==2.9.0
moviepy
source